Proibir de remover despesas que não estejam enviadas para aprovacao

+++ src/main/java/br/com/zup/sistemareembolso/services/DespesaService.java
@@ -150,6 +150,10 @@ public class DespesaService {
             throw new DespesaJaAprovadaException();
         }
 
+        if (despesa.getStatus().equals(Status.REPROVADO)) {
+            throw new DespesaJaReprovadaException();
+        }
+
         notaFiscalService.excluirNotaFiscalPeloCodigo(despesa.getNotaFiscal().getCodigoDaNota());
         despesaRepository.delete(despesa);
     }
